feat: replace version globally
Replace the version string globally, in case it is found multiple times
in the file. Some of our files reference the version multiple times,
e.g. one time in the header and another time further down as a variable.
The current version of grunt-bump only replaces the first occurrence.
Replacing the version globally can be enabled using the new
_globalReplace_ option, which by default is _false_ to keep the current
behavior as a default.