Improvement to support setting the regex per file group. Backwards compatible with the old notation.
New Notation:
files: [
                    {
                        path: 'index.html',
                        regexp: /(dist\/\w+\/\w+[-](([\d-]+\.)+[\d-]+))/g
                    },
                    {
                        path: ['package.json', 'bower.json']
                    }
                ]

Old Notation:
files: ['package.json','package.json', 'bower.json']

If no regexp property is defined within the file, the script falls back to use the regExp prop from the bump config object. If regExp prop is not defined, it uses the built-in default regexp.

Please mind, that I didn't refine the regex matching logic within the code, so you'll have to be careful with your regexes, and mind that the code expects to get the match in a strictly defined order. Use --dry-run to check if everything is ok.